Extracted DuktapeFunction into separate file
