+++ Code/Engine/Core/Scripting/DuktapeWrapper/DuktapeWrapper.cpp
@@ -335,133 +335,6 @@ void ezDuktapeWrapper::DukFree(void* pUserData, void* pPointer)
   pDukWrapper->m_Allocator.Deallocate(pPointer);
 }
 
-//////////////////////////////////////////////////////////////////////////
-//////////////////////////////////////////////////////////////////////////
-//////////////////////////////////////////////////////////////////////////
-
-ezDuktapeFunction::ezDuktapeFunction(duk_context* pExistingContext)
-  : ezDuktapeWrapper(pExistingContext)
-{
-}
-
-
-ezUInt32 ezDuktapeFunction::GetNumVarArgFunctionParameters() const
-{
-  return duk_get_top(GetContext());
-}
-
-ezInt16 ezDuktapeFunction::GetFunctionMagicValue() const
-{
-  return duk_get_current_magic(GetContext());
-}
-
-bool ezDuktapeFunction::GetBoolParameter(ezUInt32 uiArgIdx, bool fallback /*= false*/) const
-{
-  return duk_get_boolean_default(GetContext(), uiArgIdx, fallback);
-}
-
-ezInt32 ezDuktapeFunction::GetIntParameter(ezUInt32 uiArgIdx, ezInt32 fallback /*= 0*/) const
-{
-  return duk_get_int_default(GetContext(), uiArgIdx, fallback);
-}
-
-float ezDuktapeFunction::GetFloatParameter(ezUInt32 uiArgIdx, float fallback /*= 0*/) const
-{
-  return static_cast<float>(duk_get_number_default(GetContext(), uiArgIdx, fallback));
-}
-
-double ezDuktapeFunction::GetNumberParameter(ezUInt32 uiArgIdx, double fallback /*= 0*/) const
-{
-  return duk_get_number_default(GetContext(), uiArgIdx, fallback);
-}
-
-const char* ezDuktapeFunction::GetStringParameter(ezUInt32 uiArgIdx, const char* fallback /*= ""*/) const
-{
-  return duk_get_string_default(GetContext(), uiArgIdx, fallback);
-}
-
-bool ezDuktapeFunction::IsParameterOfType(ezUInt32 uiArgIdx, ezBitflags<ezDuktapeTypeMask> mask) const
-{
-  return duk_check_type_mask(GetContext(), uiArgIdx, mask.GetValue());
-}
-
-bool ezDuktapeFunction::IsParameterBool(ezUInt32 uiArgIdx) const
-{
-  return duk_check_type_mask(GetContext(), uiArgIdx, DUK_TYPE_MASK_BOOLEAN);
-}
-
-bool ezDuktapeFunction::IsParameterNumber(ezUInt32 uiArgIdx) const
-{
-  return duk_check_type_mask(GetContext(), uiArgIdx, DUK_TYPE_MASK_NUMBER);
-}
-
-bool ezDuktapeFunction::IsParameterString(ezUInt32 uiArgIdx) const
-{
-  return duk_check_type_mask(GetContext(), uiArgIdx, DUK_TYPE_MASK_STRING);
-}
-
-bool ezDuktapeFunction::IsParameterNull(ezUInt32 uiArgIdx) const
-{
-  return duk_check_type_mask(GetContext(), uiArgIdx, DUK_TYPE_MASK_NULL);
-}
-
-bool ezDuktapeFunction::IsParameterUndefined(ezUInt32 uiArgIdx) const
-{
-  return duk_check_type_mask(GetContext(), uiArgIdx, DUK_TYPE_MASK_UNDEFINED);
-}
-
-bool ezDuktapeFunction::IsParameterObject(ezUInt32 uiArgIdx) const
-{
-  return duk_check_type_mask(GetContext(), uiArgIdx, DUK_TYPE_MASK_OBJECT);
-}
-
-ezInt32 ezDuktapeFunction::ReturnVoid()
-{
-  return 0;
-}
-
-ezInt32 ezDuktapeFunction::ReturnNull()
-{
-  duk_push_null(GetContext());
-  return 1;
-}
-
-ezInt32 ezDuktapeFunction::ReturnUndefined()
-{
-  duk_push_undefined(GetContext());
-  return 1;
-}
-
-ezInt32 ezDuktapeFunction::ReturnBool(bool value)
-{
-  duk_push_boolean(GetContext(), value);
-  return 1;
-}
-
-ezInt32 ezDuktapeFunction::ReturnInt(ezInt32 value)
-{
-  duk_push_int(GetContext(), value);
-  return 1;
-}
-
-ezInt32 ezDuktapeFunction::ReturnFloat(float value)
-{
-  duk_push_number(GetContext(), value);
-  return 1;
-}
-
-ezInt32 ezDuktapeFunction::ReturnNumber(double value)
-{
-  duk_push_number(GetContext(), value);
-  return 1;
-}
-
-ezInt32 ezDuktapeFunction::ReturnString(const char* value)
-{
-  duk_push_string(GetContext(), value);
-  return 1;
-}
-
 ezResult ezDuktapeWrapper::OpenObject(const char* szObjectName)
 {
   if (m_States.m_iOpenObjects == 0)
@@ -500,3 +373,4 @@ bool ezDuktapeWrapper::HasProperty(const char* szPropertyName)
 }
 
 #endif
+
